requirements.txt
transformers
torch
librosa
soundfile
fastapi
python-multipart
httpx
uvicorn
python-dotenv
langgraph
langchain
google-genai